Requested Discount Done

+++ app/Http/Controllers/requestedDiscountController.php
@@ -5,22 +5,32 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\requestedDiscount;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Validation\UnauthorizedException;
+use Mockery\Exception;
 
 
 class requestedDiscountController extends Controller
 {
-    public function list() {
+    public function list()
+    {
         $userId = Auth::user()->id;
-        return requestedDiscount::all()->where('userId','=',$userId);
+        return requestedDiscount::query()->where('userId', '=', $userId)->get();
     }
 
-    public function get($id) {
-        // Check if id is numeric
+    public function get($id)
+    {
+        $request = new Request();
+        $request['id'] = $id;
+
+        $data = $this->validate($request, [
+            'id' => 'required|numeric'
+        ]);
         $userId = Auth::user()->id;
-        return requestedDiscount::query()->where('userId','=',$userId)->where('id','=',$id)->firstOrFail();
+        return requestedDiscount::query()->where('userId', '=', $userId)->where('id', '=', $data['id'])->first();
     }
 
-    public function post(Request $request) {
+    public function post(Request $request)
+    {
 
         $data = $this->validate($request, [
             'category' => 'required|numeric',
@@ -42,32 +52,48 @@ class requestedDiscountController extends Controller
         return $requestedDiscount;
     }
 
-    public function put($id, Request $request) {
-// Check if id is numeric
-        $userId = Auth::user()->id;
-        $requestedDiscount = requestedDiscount::query()->where('userId','=',$userId)->where('id','=',$id)->firstOrFail();
+    public function put($id, Request $request)
+    {
+        $request['id'] = $id;
 
         $data = $this->validate($request, [
+            'id' => 'required|numeric',
             'category' => 'required|numeric',
             'price' => 'required|numeric',
             'tags' => 'required|string'
         ]);
 
+        $userId = Auth::user()->id;
 
-        $requestedDiscount->category = $data['category'];
-        $requestedDiscount->price = $data['price'];
-        $requestedDiscount->tags = $data['tags'];
+        $requestedDiscount = requestedDiscount::query()->where('userId', '=', $userId)->where('id', '=', $data['id'])->first();
+        if ($requestedDiscount != null) {
+            $requestedDiscount->category = $data['category'];
+            $requestedDiscount->price = $data['price'];
+            $requestedDiscount->tags = $data['tags'];
 
 
-        $requestedDiscount->save();
-        $requestedDiscount->push();
-        return $requestedDiscount;
+            $requestedDiscount->save();
+            $requestedDiscount->push();
+            return $requestedDiscount;
+        } else {
+            return "";
+        }
     }
 
-    public function delete($id) {
-        // Check if id is numeric
+    public function delete($id)
+    {
+        $request = new Request();
+        $request['id'] = $id;
+
+        $data = $this->validate($request, [
+            'id' => 'required|numeric'
+        ]);
+
         $userId = Auth::user()->id;
-        $requestedDiscount = requestedDiscount::query()->where('userId','=',$userId)->findOrFail($id);
-        $requestedDiscount->delete();
+
+        $requestedDiscount = requestedDiscount::query()->where('userId', '=', $userId)->find($data['id']);
+        if ($requestedDiscount != null) {
+            $requestedDiscount->delete();
+        }
     }
 }
